fix sonar issues

+++ src/mock_bpa/decode.c
@@ -220,7 +220,7 @@ int bsl_mock_decode_primary(QCBORDecodeContext *dec, MockBPA_PrimaryBlock_t *blk
     const size_t end = QCBORDecode_Tell(dec);
 
     const UsefulBufC buf = QCBORDecode_RetrieveUndecodedInput(dec);
-    if (!mock_bpa_crc_check(buf, begin, end, blk->crc_type, crc_view.len))
+    if (!mock_bpa_crc_check(buf, begin, end, (int)blk->crc_type, crc_view.len))
     {
         return 4;
     }
@@ -289,7 +289,7 @@ int bsl_mock_decode_canonical(QCBORDecodeContext *dec, MockBPA_CanonicalBlock_t 
     }
 
     const UsefulBufC buf = QCBORDecode_RetrieveUndecodedInput(dec);
-    if (!mock_bpa_crc_check(buf, begin, end, blk->crc_type, crc_view.len))
+    if (!mock_bpa_crc_check(buf, begin, end, (int)blk->crc_type, crc_view.len))
     {
         return 4;
     }
